tests/test_metrics.py
import unittest
from rurage.metrics import (
    compute_nli_score,
    compute_similarity,
    compute_overlap,
    calculate_rouge,
    calculate_bleu
)

class TestMetrics(unittest.TestCase):
    def test_compute_nli_score(self):
        # Add mock test for NLI score computation
        self.assertIsNotNone(compute_nli_score(None, "context", "answer"))

    def test_compute_similarity(self):
        # Add mock test for similarity computation
        self.assertIsNotNone(compute_similarity("context", "answer"))

    def test_compute_overlap(self):
        # Add mock test for overlap computation
        self.assertIsNotNone(compute_overlap("reference", "candidate"))

    def test_calculate_rouge(self):
        # Add mock test for ROUGE calculation
        self.assertIsNotNone(calculate_rouge("reference", "candidate"))

    def test_calculate_bleu(self):
        # Add mock test for BLEU calculation
        self.assertIsNotNone(calculate_bleu("reference", "candidate"))

if __name__ == "__main__":
    unittest.main()
